+++ Servers/UI/OJS.Servers.Ui/ClientApp/package.json
@@ -16,6 +16,7 @@
     "@mui/styles": "^5.2.3",
     "@mui/system": "^5.2.4",
     "@mui/x-data-grid": "^5.2.0",
+    "@reduxjs/toolkit": "^1.9.7",
     "@types/file-saver": "^2.0.5",
     "@types/prismjs": "^1.16.6",
     "@types/react-youtube": "^7.10.0",
@@ -45,6 +46,7 @@
     "react-modal": "^3.14.3",
     "react-monaco-editor": "^0.50.1",
     "react-notifications-component": "^4.0.1",
+    "react-redux": "^8.1.3",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
