+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/guidelines/lists/List.tsx
@@ -88,7 +88,6 @@ const List = <TValue, >({
             if (isNil(values) || isEmpty(values)) {
                 return null;
             }
-
             return values.map((value) => (
                 <li key={keyFunc(value)} className={itemClassNameCombined}>
                     {itemFunc(value)}
