+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/details/SubmissionDetails.tsx
@@ -1,117 +1,44 @@
 import React, { useCallback, useEffect, useMemo } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import { useParams } from 'react-router-dom';
 import first from 'lodash/first';
 import isEmpty from 'lodash/isEmpty';
 import isNil from 'lodash/isNil';
 
-import { contestParticipationType } from '../../../common/contest-helpers';
-import { useProblemSubmissions } from '../../../hooks/submissions/use-problem-submissions';
-import { useSubmissionsDetails } from '../../../hooks/submissions/use-submissions-details';
 import { useAuth } from '../../../hooks/use-auth';
+import { IErrorDataType } from '../../../hooks/use-http';
 import { usePageTitles } from '../../../hooks/use-page-titles';
-import { usePages } from '../../../hooks/use-pages';
-import { useProblems } from '../../../hooks/use-problems';
+import { setCurrentSubmissionResults, setSubmission } from '../../../redux/features/submissionDetailsSlice';
+import { useGetCurrentSubmissionQuery, useGetSubmissionResultsQuery } from '../../../redux/services/submissionDetailsService';
 import concatClassNames from '../../../utils/class-names';
-import { preciseFormatDate } from '../../../utils/dates';
 import { flexCenterObjectStyles } from '../../../utils/object-utils';
-import { getAdministrationRetestSubmissionInternalUrl, getParticipateInContestUrl } from '../../../utils/urls';
-import CodeEditor from '../../code-editor/CodeEditor';
-import AlertBox, { AlertBoxType } from '../../guidelines/alert-box/AlertBox';
-import { Button, ButtonSize, ButtonState, ButtonType, LinkButton, LinkButtonType } from '../../guidelines/buttons/Button';
+import { getAdministrationRetestSubmissionInternalUrl } from '../../../utils/urls';
+import { ButtonSize, LinkButton, LinkButtonType } from '../../guidelines/buttons/Button';
 import Heading, { HeadingType } from '../../guidelines/headings/Heading';
-import IconSize from '../../guidelines/icons/common/icon-sizes';
-import LeftArrowIcon from '../../guidelines/icons/LeftArrowIcon';
-import PaginationControls from '../../guidelines/pagination/PaginationControls';
 import SpinningLoader from '../../guidelines/spinning-loader/SpinningLoader';
 import SubmissionResults from '../submission-results/SubmissionResults';
-import SubmissionsList from '../submissions-list/SubmissionsList';
 
+import RefreshableSubmissionList from './refreshable-submission-list/RefreshableSubmissionList';
 import SubmissionResultsDetails from './submission-result-details/SubmissionResultsDetails';
+import SubmissionDetailsCodeEditor from './submissionDetails-codeEditor/SubmissionDetailsCodeEditor';
 
 import styles from './SubmissionDetails.module.scss';
 
 const SubmissionDetails = () => {
-    const {
-        state: {
-            isLoading,
-            currentSubmission,
-            currentSubmissionResults,
-            validationErrors,
-            downloadErrorMessage,
-        },
-        actions: {
-            getResults,
-            downloadProblemSubmissionFile,
-            setDownloadErrorMessage,
-            setCurrentSubmission,
-            selectSubmissionById,
-            setSubmissionResultsUrlParams,
-            getDetails,
-        },
-    } = useSubmissionsDetails();
-    const {
-        state: { problemSubmissionsPage },
-        actions: { changeProblemSubmissionsPage },
-    } = useProblemSubmissions();
-    const { actions: { initiateRedirectionToProblem } } = useProblems();
     const { actions: { setPageTitle } } = usePageTitles();
     const { state: { user: { permissions: { canAccessAdministration } } } } = useAuth();
+    const dispatch = useDispatch();
+    const { currentSubmission, validationErrors, currentPage } =
+    useSelector((state: any) => state.submissionDetails);
+    const { submissionId } = useParams();
+    const { data: currentSubmissionData, isFetching, refetch } = useGetCurrentSubmissionQuery({ submissionId: Number(submissionId) });
+    const { data, isFetching: isLoadingResults, refetch: refetchResults } =
+    useGetSubmissionResultsQuery({ submissionId: Number(submissionId), page: currentPage });
 
-    const { state: { user } } = useAuth();
-    const { state: { pagesInfo } } = usePages();
-
-    const renderDownloadErrorMessage = useCallback(() => {
-        if (isNil(downloadErrorMessage)) {
-            return null;
-        }
-
-        return (
-            <AlertBox
-              message={downloadErrorMessage}
-              type={AlertBoxType.error}
-              onClose={() => setDownloadErrorMessage(null)}
-            />
-        );
-    }, [ downloadErrorMessage, setDownloadErrorMessage ]);
-
-    const handleDownloadSubmissionFile = useCallback(
-        async () => {
-            if (isNil(currentSubmission)) {
-                return;
-            }
-
-            const { id } = currentSubmission;
-            await downloadProblemSubmissionFile(id);
-        },
-        [ downloadProblemSubmissionFile, currentSubmission ],
-    );
-
-    const renderResourceLink = useCallback(
-        () => {
-            if (isNil(currentSubmission)) {
-                return null;
-            }
-
-            const { submissionType: { allowBinaryFilesUpload }, user: { userName: submissionUserName } } = currentSubmission;
-            const { username: loggedInUserName } = user;
-            if ((!canAccessAdministration && submissionUserName !== loggedInUserName) || !allowBinaryFilesUpload) {
-                return null;
-            }
-
-            return (
-                <div className={styles.resourceWrapper}>
-                    <Button
-                      type={ButtonType.primary}
-                      className={styles.resourceLinkButton}
-                      onClick={() => handleDownloadSubmissionFile()}
-                    >
-                        Download file
-                    </Button>
-                </div>
-            );
-        },
-        [ handleDownloadSubmissionFile, canAccessAdministration, currentSubmission, user ],
-    );
-
+    useEffect(() => {
+        dispatch(setSubmission(currentSubmissionData));
+        dispatch(setCurrentSubmissionResults(data));
+    }, [ currentSubmissionData, data, dispatch ]);
     useEffect(
         () => {
             if (currentSubmission) {
@@ -120,17 +47,10 @@ const SubmissionDetails = () => {
         },
         [ currentSubmission, setPageTitle ],
     );
-
-    const problemNameHeadingText = useMemo(
-        () => {
-            if (!currentSubmission) {
-                return '';
-            }
-
-            return `${currentSubmission?.problem.name} - ${currentSubmission?.problem.id}`;
-        },
-        [ currentSubmission ],
-    );
+    const reloadPage = useCallback(() => {
+        refetch();
+        refetchResults();
+    }, [ refetch, refetchResults ]);
 
     const detailsHeadingText = useMemo(
         () => (
@@ -142,42 +62,11 @@ const SubmissionDetails = () => {
         [ currentSubmission?.id ],
     );
 
-    const { submissionType } = currentSubmission || {};
-
     const submissionsDetails = 'submissionDetails';
     const submissionDetailsClassName = concatClassNames(
         styles.navigation,
         styles.submissionDetails,
         submissionsDetails,
-    );
-
-    const handleReloadClick = useCallback(
-        async () => {
-            if (isNil(currentSubmission)) {
-                return;
-            }
-
-            const { id: submissionId } = currentSubmission;
-
-            setSubmissionResultsUrlParams({
-                submissionId,
-                page: problemSubmissionsPage,
-            });
-
-            getDetails(submissionId);
-        },
-        [ problemSubmissionsPage, currentSubmission, setSubmissionResultsUrlParams, getDetails ],
-    );
-
-    const handlePageChange = useCallback(
-        (page: number) => {
-            changeProblemSubmissionsPage(page);
-
-            const { id } = currentSubmission!;
-
-            getResults(id, page);
-        },
-        [ changeProblemSubmissionsPage, getResults, currentSubmission ],
     );
 
     const renderRetestButton = useCallback(
@@ -199,210 +88,10 @@ const SubmissionDetails = () => {
         },
         [ canAccessAdministration, currentSubmission ],
     );
-    const renderButtonsSection = useCallback(() => (
-        <div className={styles.buttonsSection}>
-            <Button
-              onClick={handleReloadClick}
-              text="Reload"
-              type={ButtonType.secondary}
-              className={styles.submissionReloadBtn}
-            />
-            {renderRetestButton()}
-        </div>
-    ), [ handleReloadClick, renderRetestButton ]);
 
-    const renderTestsChangeMessage = useCallback(() => (
-        currentSubmission?.testRuns.length === 0 &&
-        currentSubmission.isCompiledSuccessfully &&
-        currentSubmission.totalTests > 0 &&
-        !currentSubmission.processingComment &&
-        currentSubmission.isProcessed
-            ? (
-                <div className={styles.testChangesWrapper}>
-                    <p>
-                        The input/output data changed. Your (
-                        {currentSubmission.points}
-                        /
-                        {currentSubmission.problem.maximumPoints}
-                        )
-                        submission is now outdated.
-                        Click &quot;Retest&quot; to resubmit your solution for re-evaluation against the new test cases.
-                        Your score may change.
-                    </p>
-                    {renderRetestButton()}
-                </div>
-            )
-            : ''
-    ), [ currentSubmission, renderRetestButton ]);
-
-    const renderSubmissionInfo = useCallback(
-        () => {
-            if (!canAccessAdministration || isNil(currentSubmission)) {
-                return null;
-            }
-
-            const { createdOn, modifiedOn, startedExecutionOn, completedExecutionOn, user: { userName } } = currentSubmission;
-
-            return (
-                <div className={styles.submissionInfo}>
-                    <p className={styles.submissionInfoParagraph}>
-                        Created on:
-                        {' '}
-                        {preciseFormatDate(createdOn)}
-                    </p>
-                    <p className={styles.submissionInfoParagraph}>
-                        Modified on:
-                        {' '}
-                        {isNil(modifiedOn)
-                            ? 'never'
-                            : preciseFormatDate(modifiedOn)}
-                    </p>
-                    <p className={styles.submissionInfoParagraph}>
-                        Started execution on:
-                        {' '}
-                        {isNil(startedExecutionOn)
-                            ? 'never'
-                            : preciseFormatDate(startedExecutionOn)}
-                    </p>
-                    <p className={styles.submissionInfoParagraph}>
-                        Completed execution on:
-                        {' '}
-                        {isNil(completedExecutionOn)
-                            ? 'never'
-                            : preciseFormatDate(completedExecutionOn)}
-                    </p>
-                    <p className={styles.submissionInfoParagraph}>
-                        Username:
-                        {' '}
-                        {userName}
-                    </p>
-                </div>
-            );
-        },
-        [ currentSubmission, canAccessAdministration ],
-    );
-
-    const backButtonState = useMemo(
-        () => isNil(currentSubmission?.contestId)
-            ? ButtonState.disabled
-            : ButtonState.enabled,
-        [ currentSubmission ],
-    );
-
-    const { pagesCount } = pagesInfo;
-    const refreshableSubmissionsList = useCallback(
-        () => (
-            <div className={styles.navigation}>
-                <div style={{ marginBottom: '24px' }}>
-                    <Heading type={HeadingType.secondary}>Submissions</Heading>
-                </div>
-                <SubmissionsList
-                  items={currentSubmissionResults}
-                  selectedSubmission={currentSubmission}
-                  className={styles.submissionsList}
-                />
-                <PaginationControls
-                  count={pagesCount}
-                  page={problemSubmissionsPage}
-                  onChange={handlePageChange}
-                />
-                { renderButtonsSection() }
-                { renderSubmissionInfo() }
-            </div>
-        ),
-        [ currentSubmissionResults, currentSubmission, pagesCount,
-            problemSubmissionsPage, handlePageChange, renderButtonsSection, renderSubmissionInfo ],
-    );
-
-    const setSubmissionAndStartParticipation = useCallback(
-        (contestId: number) => {
-            const participationType = contestParticipationType(currentSubmission!.isOfficial);
-
-            const participateInContestUrl = getParticipateInContestUrl({
-                id: contestId,
-                participationType,
-            });
-
-            const { problem: { id: problemId } } = currentSubmission!;
-
-            initiateRedirectionToProblem(problemId, participateInContestUrl);
-
-            setCurrentSubmission(null);
-            selectSubmissionById(null);
-        },
-        [ currentSubmission, selectSubmissionById, setCurrentSubmission, initiateRedirectionToProblem ],
-    );
-
-    const codeEditor = useCallback(
-        () => (
-            <div className={styles.code}>
-                <Heading
-                  type={HeadingType.secondary}
-                  className={styles.taskHeading}
-                  style={!isNil(problemNameHeadingText) && problemNameHeadingText.length >= 30
-                      ? { marginBottom: 0 }
-                      : { marginBottom: '24px' }}
-                >
-                    <div className={styles.btnContainer}>
-                        <LeftArrowIcon className={styles.leftArrow} size={IconSize.Large} />
-                        <Button
-                          type={ButtonType.secondary}
-                          size={ButtonSize.small}
-                          onClick={() => setSubmissionAndStartParticipation(currentSubmission!.contestId)}
-                          className={styles.backBtn}
-                          text=" "
-                          state={backButtonState}
-                        />
-                    </div>
-                    <div style={{ maxWidth: '30ch', textAlign: 'center' }}>
-                        {problemNameHeadingText}
-                    </div>
-                </Heading>
-                <div>
-                    {renderTestsChangeMessage()}
-                </div>
-                {
-                        !currentSubmission?.isProcessed
-                            ? (
-                                <AlertBox
-                                  className={styles.alertBox}
-                                  message="The submission is in queue and will be processed shortly. Please wait."
-                                  type={AlertBoxType.info}
-                                  isClosable={false}
-                                />
-                            )
-                            : null
-                    }
-                {submissionType?.allowBinaryFilesUpload
-                    ? (
-                        <div className={styles.resourceWrapper}>
-                            {renderResourceLink()}
-                            {renderDownloadErrorMessage()}
-                        </div>
-                    )
-                    : (
-                        <CodeEditor
-                          readOnly
-                          code={currentSubmission?.content}
-                          selectedSubmissionType={submissionType}
-                        />
-                    )}
-            </div>
-        ),
-        [
-            problemNameHeadingText,
-            submissionType,
-            backButtonState,
-            renderResourceLink,
-            renderDownloadErrorMessage,
-            currentSubmission,
-            setSubmissionAndStartParticipation,
-            renderTestsChangeMessage,
-        ],
-    );
-
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
     const submissionResults = useCallback(
-        () => (isLoading
+        () => (isFetching
             ? (
                 <div style={{ ...flexCenterObjectStyles }}>
                     <SpinningLoader />
@@ -422,12 +111,12 @@ const SubmissionDetails = () => {
                         )}
                 </div>
             )),
-        [ currentSubmission, detailsHeadingText, submissionDetailsClassName, isLoading ],
+        [ currentSubmission, detailsHeadingText, submissionDetailsClassName, isFetching ],
     );
 
     const renderErrorMessage = useCallback(
         () => {
-            const error = first(validationErrors);
+            const error = first(validationErrors as IErrorDataType[]);
             if (!isNil(error)) {
                 const { detail } = error;
                 return (
@@ -444,7 +133,7 @@ const SubmissionDetails = () => {
         [ validationErrors ],
     );
 
-    if (!isLoading && isNil(currentSubmission) && isEmpty(validationErrors)) {
+    if (!isFetching && isNil(currentSubmission) && isEmpty(validationErrors)) {
         return <div>No details fetched.</div>;
     }
 
@@ -452,7 +141,7 @@ const SubmissionDetails = () => {
         return renderErrorMessage();
     }
 
-    if (isLoading) {
+    if (isFetching || isLoadingResults) {
         return (
             <div style={{ ...flexCenterObjectStyles }}>
                 <SpinningLoader />
@@ -463,8 +152,11 @@ const SubmissionDetails = () => {
     return (
         <>
             <div className={styles.detailsWrapper}>
-                {refreshableSubmissionsList()}
-                {codeEditor()}
+                <RefreshableSubmissionList
+                  renderRetestButton={renderRetestButton}
+                  reload={reloadPage}
+                />
+                <SubmissionDetailsCodeEditor renderRetestButton={renderRetestButton} />
                 {submissionResults()}
             </div>
             <SubmissionResultsDetails testRuns={currentSubmission?.testRuns} />
