+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/details/refreshable-submission-list/RefreshableSubmissionList.module.scss
@@ -1,0 +1,57 @@
+@use 'src/styles/global';
+@use 'src/styles/components';
+@use 'src/styles/fonts';
+@use 'src/styles/variables';
+@use 'src/styles/colors';
+
+.buttonsSection {
+    margin-top: 20px;
+}
+
+.submissionInfo {
+    font-family: fonts.$font-family-lato-regular;
+    margin-top: 15px;
+}
+
+.submissionInfoParagraph {
+    @extend %font-family-lato;
+  
+    color: colors.$text-dark-gray-color;
+    margin-bottom: 10px;
+}
+
+.navigation,
+%navigation {
+  align-content: center;
+  flex: 3;
+}
+
+.submissionsList {
+    height: 465px;
+  
+    ul {
+      @extend %customScroll;
+  
+      overflow-x: hidden;
+      overflow-y: auto;
+    }
+}
+
+.submissionReloadBtn {
+    margin-right: 20px;
+    margin-top: 10px;
+}
+
+@media only screen and (max-width: 1430px) {
+    .buttonsSection {
+      display: flex;
+      flex-direction: column;
+      justify-content: flex-start;
+  
+      a {
+        margin-left: 0;
+        margin-right: 20px;
+        margin-top: 10px;
+      }
+    }
+  }
