+++ Servers/UI/OJS.Servers.Ui/ClientApp/src/components/submissions/details/submissionDetails-codeEditor/SubmissionDetailsCodeEditor.module.scss
@@ -1,0 +1,79 @@
+@use 'src/styles/global';
+@use 'src/styles/components';
+@use 'src/styles/fonts';
+@use 'src/styles/variables';
+@use 'src/styles/colors';
+
+.resourceWrapper {
+    align-items: center;
+    display: flex;
+    flex-direction: row;
+    gap: 10px;
+    min-height: 50px;
+}
+
+.resourceLinkButton {
+    text-decoration: none;
+}
+
+.testChangesWrapper {
+    align-items: center;
+    background-color: colors.$warning-color;
+    border: 2px solid colors.$warning-color;
+    border-radius: 12px;
+    color: colors.$white-color;
+    display: flex;
+    font-family: fonts.$font-family-lato;
+    font-size: 16px;
+    justify-content: space-around;
+    margin-bottom: 1rem;
+    padding: 1rem;
+}
+
+.code {
+    align-content: center;
+    flex: 3;
+  
+    border-left: 0;
+    flex: 7;
+  
+    @media only screen and (max-width: variables.$r-large-desktop-max) {
+      flex: 3;
+    }
+}
+
+.taskHeading {
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+  
+    > div {
+      > div {
+        pointer-events: none;
+      }
+  
+      button {
+        margin-right: 30px;
+      }
+    }
+}
+
+.btnContainer {
+    display: flex;
+    flex-direction: row;
+}
+
+.leftArrow {
+    color: colors.$light-blue-color;
+    margin-bottom: -5px;
+}
+
+.backBtn {
+    border: 0;
+    margin-left: -30px;
+    margin-right: -30px;
+}
+
+.alertBox {
+    margin-bottom: 20px;
+  }
